Last refactoring steps
