+++ core/src/main/java/com/g16/feyrune/interfaces/IAbility.java
@@ -1,6 +1,6 @@
 package com.g16.feyrune.interfaces;
 
-public interface IMove {
+public interface IAbility {
     int getAttackPower();
     int getAttackAccuracy();
     String getAttackName();
