+++ core/src/main/java/com/g16/feyrune/interfaces/ICreature.java
@@ -3,7 +3,7 @@ package com.g16.feyrune.interfaces;
 import java.util.List;
 
 public interface ICreature {
-    List<IMove> getMoves();
+    List<IAbility> getMoves();
     double getHP();
     int getSpeed();
     int getPower();
