+++ core/src/main/java/com/g16/feyrune/model/combat/AttackHandler.java
@@ -1,7 +1,7 @@
 package com.g16.feyrune.model.combat;
 
 import com.g16.feyrune.interfaces.ICombatCreature;
-import com.g16.feyrune.interfaces.IMove;
+import com.g16.feyrune.interfaces.IAbility;
 
 import com.g16.feyrune.Util.Random;
 
@@ -13,7 +13,7 @@ public class AttackHandler {
      * @param defender The monster trying to evade the attack
      * @param attack The specific attack being used
      */
-    public static void handleAttack(ICombatCreature attacker, ICombatCreature defender, IMove attack){
+    public static void handleAttack(ICombatCreature attacker, ICombatCreature defender, IAbility attack){
         if (evasiveManoeuvre(defender, attack)){
             int damage = calculateDamage(attacker, attack);
             defender.takeDamage(damage);
@@ -26,7 +26,7 @@ public class AttackHandler {
      * @param attack: the attack action used against the defender
      * @return damage dealt to the defender (only calculates, doesn't deal it)
      */
-    private static int calculateDamage(ICombatCreature attacker, IMove attack){
+    private static int calculateDamage(ICombatCreature attacker, IAbility attack){
         int damage = attacker.getPower() * attack.getAttackPower();
         return damage;
     }
@@ -37,7 +37,7 @@ public class AttackHandler {
      * @param baseAttack: attack to check accuracy on.
      * @return if attack is successful: true, if it misses false
      */
-    private static boolean evasiveManoeuvre(ICombatCreature defender, IMove baseAttack){
+    private static boolean evasiveManoeuvre(ICombatCreature defender, IAbility baseAttack){
         boolean evasion = defender.getSpeed()< Random.randomInt(100);
         boolean accuracy = baseAttack.getAttackAccuracy()>Random.randomInt(100);
         boolean hit= evasion && accuracy;
