+++ core/src/main/java/com/g16/feyrune/model/creature/CreatureFactory.java
@@ -1,8 +1,7 @@
 package com.g16.feyrune.model.creature;
 
 import com.g16.feyrune.interfaces.ICreature;
-import com.g16.feyrune.model.action.BaseAttack;
-import sun.reflect.generics.reflectiveObjects.NotImplementedException;
+import com.g16.feyrune.model.combat.actions.abilities.BaseAbility;
 import com.g16.feyrune.Util.Pair;
 
 public class CreatureFactory {
@@ -11,9 +10,9 @@ public class CreatureFactory {
      * @return a new Object of type ICreature
      */
     public static ICreature createCreature() {
-        BaseAttack[] baseAttackList= new BaseAttack[1];
-        baseAttackList[0] = new BaseAttack(100,100,"SuperAwesomeBaseAttack");
-        return new BaseCreature(100,75,50,33, baseAttackList);
+        BaseAbility[] baseAbilityList = new BaseAbility[1];
+        baseAbilityList[0] = new BaseAbility(100,100,"SuperAwesomeBaseAttack");
+        return new BaseCreature(100,75,50,33, baseAbilityList);
 
     }
     public static ICreature[] createCreatureList(Pair<Integer, Integer>[] monsterList){
